pyproject.toml
[project]
name = "py-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.5.0",
]
